+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - perf - improve performance by hoisting inlined regexps out of loops and subfunctions
 
 # v2022.5.1-beta
+- bugfix - fix jslint falsely believing megastring literals `0` and `1` are similar
 - bugfix - fix function jstestOnExit() from exiting prematurely and suppressing additional error-messages
 
 # v2022.3.30